+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 	* sanity checks for manually provided public and private keys
 	* handle protocol changes more elegantly
 	* more sophisticated version checking
+	* removed obsolete `carefulness` CLI option
 * v9.47.1 (2014-12-02)
 	* change first testnet block to 310000
 * v9.47.0 (2014-12-01)